+++ src/models/KeyValueI.ts
@@ -1,0 +1,3 @@
+export interface KeyValueI {
+  [key: string]: any
+}