+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- additional verification for token: 
+  - verify token age
+  - verify sso params
+  - website signature
+- support for migration
+- `jwt` and `queryParams` config moved from toolbox
+
+### Changed
+- all SSO params have prefix `sso`
+
+### Removed
+- `refId` is not longer part of token
+
 ## [0.4.4] - 2020-12-09
 ### Fixed
 - fix invalid publish of dist/models/JwtPayload file
@@ -25,7 +40,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - rewrite code to TypeScript
 
 ### Fixed
-- fix test for `fetchSilkeyPublicKey()`
+- fix test for `fetchSilkeyEthAddress()`
 
 ### Removed
 - remove `isEmpty` method from utils